encoder: the chroma/bit-depth axis — every §6.4.1 (depth, format) pairing encodes
The conformance-grade KEY / inter-GOP / pyramid / adaptive encoders
stop being 8-bit-4:2:0-only. New general entries
encode_key_frame_yuv{,_with_q}, encode_gop_yuv{,_with_q},
encode_pyramid_gop_yuv_with_q and encode_adaptive_gop_yuv_with_q take
a YuvFrame (u16 planes + bit_depth + ChromaFormat) at 8/10/12-bit x
4:2:0/4:2:2/4:4:4/monochrome, with the §6.4.1 seq_profile election
(0 = 8/10-bit 4:2:0-or-mono, 1 = 8/10-bit 4:4:4, 2 = 8/10-bit 4:2:2
or 12-bit anything) and per-pairing §5.5.2 color_config synthesis in
the new encoder::yuv_frame module.

The whole encoder pixel pipeline moves to one u16 sample type at the
stream depth: ReconState + reference stores + motion search +
global-motion prepass + dv-hash planes, Clip1(BitDepth) recon clamps,
§7.11.2.2 depth-scaled neutral fills, §7.12.2 per-depth quantiser
rows via QuantizerParams.bit_depth, and a depth-scaled lambda
(x 2^(2(bd-8)); 8-bit decisions bit-identical). Format threading is
spec-complete: the §5.11.5 HasChroma general rule, §5.11.38
get_plane_residual_size / get_tx_size at real (subsampling_x,
subsampling_y), the general §7.11.5 CfL build ((1+subY)x(1+subX)
luma sum, t << (3-subX-subY), MaxLuma clamps at (1<<sub)), the
§5.11.40 chroma-inheritance lift at (2+subX, 2+subY), and a
monochrome walk that skips every chroma pass (NumPlanes = 1).

§5.11.38 partition admissibility lands with it: under 4:2:2 the RD
ladders (KEY rect candidates, edge split_or_* arms, P-frame
HORZ/VERT/T-shapes/strips) never elect a subSize whose chroma
residual size is BLOCK_INVALID (UV aspect ratios stay within
1:4..4:1) — surfaced by the write mirror correctly rejecting a
32x64 CfL commitment whose search-side gate had been derived off a
masked BLOCK_INVALID.

The historical *_yuv420 entries keep their exact signatures and
route through the same core byte-identically (pinned by test). The
§5.11.46 palette and §5.9.20 intra-block-copy elections stay
8-bit-4:2:0-scoped (an encoder choice — the header gates close on
the other pairings).

tests/format_matrix_conformance.rs: KEY (lossless + lossy) and
3-frame inter GOP round trips through the in-tree spec driver for
all twelve pairings (recon == decode sample-exact, lossless == input
exact, emitted profile/color-config asserted), the legacy-bytes
identity pin, and the env-gated 24-stream matrix dump for external
black-box validation (24/24 byte-identical on each of the three
independent reference decoders this round).

+++ src/encoder/dv_hash.rs
@@ -72,7 +72,7 @@ const FNV_BASIS: u64 = 0xcbf2_9ce4_8422_2325;
 const FNV_PRIME: u64 = 0x0000_0100_0000_01B3;
 
 #[inline]
-fn fnv_mix(mut hash: u64, v: u8) -> u64 {
+fn fnv_mix(mut hash: u64, v: u16) -> u64 {
     hash ^= u64::from(v);
     hash = hash.wrapping_mul(FNV_PRIME);
     hash
@@ -80,7 +80,7 @@ fn fnv_mix(mut hash: u64, v: u8) -> u64 {
 
 /// FNV-1a over an 8×8 raster block. Returns `(hash, uniform)`.
 #[inline]
-fn hash8(plane: &[u8], stride: usize, y: usize, x: usize) -> (u64, bool) {
+fn hash8(plane: &[u16], stride: usize, y: usize, x: usize) -> (u64, bool) {
     let first = plane[y * stride + x];
     let mut hash = FNV_BASIS;
     let mut uniform = true;
@@ -105,7 +105,7 @@ fn compose4(q: [(u64, bool); 4]) -> (u64, bool) {
     for &(qh, qu) in &q {
         uniform &= qu && qh == q[0].0;
         for b in qh.to_le_bytes() {
-            hash = fnv_mix(hash, b);
+            hash = fnv_mix(hash, u16::from(b));
         }
     }
     (hash, uniform)
@@ -116,7 +116,7 @@ fn compose4(q: [(u64, bool); 4]) -> (u64, bool) {
 /// `size` MUST be in [`DV_HASH_SIZES`] and the block fully inside the
 /// plane. Returns `(hash, uniform)`.
 pub(crate) fn hash_block_direct(
-    plane: &[u8],
+    plane: &[u16],
     stride: usize,
     y: usize,
     x: usize,
@@ -153,7 +153,7 @@ pub(crate) struct DvHashIndex {
 impl DvHashIndex {
     /// Build the index over a `width`×`height` input luma plane.
     /// Falls back to the inert state above [`MAX_INDEXED_AREA`].
-    pub(crate) fn build(plane: &[u8], width: usize, height: usize) -> Self {
+    pub(crate) fn build(plane: &[u16], width: usize, height: usize) -> Self {
         let mut out = DvHashIndex::default();
         if width < 8 || height < 8 || width * height > MAX_INDEXED_AREA {
             return out;
@@ -242,7 +242,7 @@ impl DvHashIndex {
 mod tests {
     use super::*;
 
-    fn pattern_plane(w: usize, h: usize, seed: u32) -> Vec<u8> {
+    fn pattern_plane(w: usize, h: usize, seed: u32) -> Vec<u16> {
         let mut state = seed | 1;
         let mut next = move || {
             state ^= state << 13;
@@ -250,7 +250,7 @@ mod tests {
             state ^= state << 5;
             state
         };
-        (0..w * h).map(|_| (next() & 0xff) as u8).collect()
+        (0..w * h).map(|_| (next() & 0xff) as u16).collect()
     }
 
     /// The grid-composed build and the direct probe recursion must
@@ -291,14 +291,14 @@ mod tests {
     #[test]
     fn uniform_suppression() {
         let (w, h) = (128, 128);
-        let mut plane = vec![77u8; w * h];
+        let mut plane = vec![77u16; w * h];
         let (_, uniform) = hash_block_direct(&plane, w, 0, 0, 64);
         assert!(uniform);
         plane[63 * w + 63] = 78;
         let (_, uniform) = hash_block_direct(&plane, w, 0, 0, 64);
         assert!(!uniform);
         // A fully-flat frame indexes nothing.
-        let flat = vec![13u8; w * h];
+        let flat = vec![13u16; w * h];
         let idx = DvHashIndex::build(&flat, w, h);
         assert!(idx.is_empty());
     }
@@ -308,7 +308,7 @@ mod tests {
     #[test]
     fn off_stride_positions_indexed() {
         let (w, h) = (256, 128);
-        let mut plane = vec![0u8; w * h];
+        let mut plane = vec![0u16; w * h];
         let glyph = pattern_plane(8, 8, 0x1357);
         for i in 0..8 {
             for j in 0..8 {
@@ -328,10 +328,10 @@ mod tests {
     fn bucket_cap_and_nearest_order() {
         let (w, h) = (448, 192);
         // Tile a non-uniform 8x8 glyph everywhere: one giant bucket.
-        let mut plane = vec![0u8; w * h];
+        let mut plane = vec![0u16; w * h];
         for y in 0..h {
             for x in 0..w {
-                plane[y * w + x] = ((x % 8) * 8 + (y % 8)) as u8;
+                plane[y * w + x] = ((x % 8) * 8 + (y % 8)) as u16;
             }
         }
         let idx = DvHashIndex::build(&plane, w, h);
